# docs/api/openapi.runtime.v1alpha1.json
{
  "components": {
    "schemas": {
...
      "CancelRunResponse": {
        "additionalProperties": false,
        "description": "Acknowledgement of a cooperative cancellation request.",
        "properties": {
          "cancellation_requested": {
            "default": true,
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": "axcalib.api-cancel/v1alpha1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineRunStatus"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "CancelRunResponse",
        "type": "object"
      },
...
      "PipelineCatalogResponse": {
        "additionalProperties": false,
        "description": "Deterministic catalog of allowlisted local pipelines.",
        "properties": {
          "pipelines": {
            "items": {
              "$ref": "#/components/schemas/PipelineDescriptor"
            },
            "title": "Pipelines",
            "type": "array"
          },
          "schema_version": {
            "default": "axcalib.api-pipeline-catalog/v1alpha1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "pipelines"
        ],
        "title": "PipelineCatalogResponse",
        "type": "object"
      },
      "PipelineDescriptor": {
        "additionalProperties": false,
        "description": "Small serializable catalog entry for delivery adapters.",
        "properties": {
          "pipeline_id": {
            "title": "Pipeline Id",
            "type": "string"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "request_type": {
            "title": "Request Type",
            "type": "string"
          },
          "result_type": {
            "title": "Result Type",
            "type": "string"
          }
        },
        "required": [
          "pipeline_id",
          "pipeline_version",
          "request_type",
          "result_type"
        ],
        "title": "PipelineDescriptor",
        "type": "object"
      },
      "PipelineRunRequest": {
        "additionalProperties": false,
        "description": "Transport options plus one pipeline-specific JSON payload.",
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "run_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "payload"
        ],
        "title": "PipelineRunRequest",
        "type": "object"
      },
...
      "PipelineRunView": {
        "additionalProperties": false,
        "description": "Filesystem-neutral execution result safe for HTTP delivery.",
        "properties": {
          "attempt": {
            "minimum": 0.0,
            "title": "Attempt",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "pipeline_id": {
            "title": "Pipeline Id",
            "type": "string"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": "axcalib.api-pipeline-run/v1alpha1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineRunStatus"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "run_id",
          "pipeline_id",
          "pipeline_version",
          "status",
          "attempt"
        ],
        "title": "PipelineRunView",
        "type": "object"
      },
      "Problem": {
        "additionalProperties": false,
        "description": "RFC 9457-shaped error body with AXCalib-stable machine codes.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ValidationIssue": {
        "additionalProperties": false,
        "description": "Redacted validation location and code without rejected input values.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          }
        },
        "required": [
          "location",
          "code"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Authenticated HTTP parity for allowlisted AXCalib pipelines",
    "title": "AXCalib Runtime API",
    "version": "v1alpha1"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/pipelines": {
      "get": {
        "operationId": "listPipelines",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Pipelines"
      }
    },
    "/v1/pipelines/{pipeline_id}/versions/{pipeline_version}/runs": {
      "post": {
        "operationId": "runPipeline",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Pipeline Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pipeline_version",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Pipeline Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Run Pipeline"
      }
    },
...
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "getRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Run"
      }
    },
    "/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancelRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Structured AXCalib API problem"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Cancel Run"
      }
    }
  },
  "servers": [
    {
      "description": "Deployment-relative base URL",
      "url": "/"
    }
  ],
...
}